requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
pydantic[email]
python-dotenv
bcrypt
python-jose[cryptography]
python-multipart
